+++ NEWS.md
@@ -1,5 +1,7 @@
 # googlesheets4 (development version)
 
+* `sheets_create()` is a new function to create a new Sheet and, optionally, write one or more data frames into it (#61). *caution: function still under development*
+
 # googlesheets4 0.1.0
 
 * Added a `NEWS.md` file to track changes to the package.